% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.ready4pack_manifest}
\alias{author.ready4pack_manifest}
\alias{author,ready4pack_manifest-method}
\title{author - a method that authors and saves files to local or remote locations}
\usage{
author.ready4pack_manifest(x)

\S4method{author}{ready4pack_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class Manifest for packages containing datasets.}
}
\value{
x (An object)
}
\description{
author.ready4pack_manifest() is an author method that authors and saves files to local or remote locations. This method is implemented for the ready4 s3 class Manifest for packages containing datasets. The function is called for its side effects and does not return a value.
}
