% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData.ready4pack_manifest}
\alias{authorData.ready4pack_manifest}
\alias{authorData,ready4pack_manifest-method}
\title{AuthorData - a method that authors and saves files necessary for creating and documenting datasets.}
\usage{
authorData.ready4pack_manifest(x)

\S4method{authorData}{ready4pack_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class Manifest for packages containing datasets.}
}
\description{
authorData.ready4pack_manifest() is an AuthorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 s3 class Manifest for packages containing datasets. The function is called for its side effects and does not return a value.
}
