% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4pack_manifest.R
\name{is_ready4pack_manifest}
\alias{is_ready4pack_manifest}
\title{is ready4pack manifest ready4 s3 class Manifest for packages containing datasets.}
\usage{
is_ready4pack_manifest(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 s3 class Manifest for packages containing datasets.
}
\description{
Check whether an object is a valid instance of the ready4 s3 class Manifest for packages containing datasets.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
