% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4pack_manifest.R
\name{make_new_ready4pack_manifest}
\alias{make_new_ready4pack_manifest}
\title{make new ready4pack manifest ready4 s3 class Manifest for packages containing datasets.}
\usage{
make_new_ready4pack_manifest(x)
}
\arguments{
\item{x}{A prototype for the ready4 s3 class Manifest for packages containing datasets.}
}
\value{
An unvalidated instance of the ready4 s3 class Manifest for packages containing datasets.
}
\description{
Create a new unvalidated instance of the ready4 s3 class Manifest for packages containing datasets.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
