% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4pack_manifest.R
\name{make_pt_ready4pack_manifest}
\alias{make_pt_ready4pack_manifest}
\title{Make prototype ready4pack manifest ready4 s3 class Manifest for packages containing datasets.}
\usage{
make_pt_ready4pack_manifest(
  x_ready4fun_manifest = ready4fun::ready4fun_manifest(),
  constructor_r3 = ready4class::ready4class_constructor(),
  pkg_ds_ls_ls = list(),
  clss_to_apply_ls = list()
)
}
\arguments{
\item{x_ready4fun_manifest}{PARAM_DESCRIPTION, Default: ready4fun::ready4fun_manifest()}

\item{constructor_r3}{Constructor (a ready4 S3), Default: ready4class::ready4class_constructor()}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists), Default: list()}

\item{clss_to_apply_ls}{Classes to apply (a list), Default: list()}
}
\value{
A prototype for ready4 s3 class Manifest for packages containing datasets.
}
\description{
Create a new prototype for the ready4 s3 class Manifest for packages containing datasets.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
