% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4pack_manifest.R
\name{validate_ready4pack_manifest}
\alias{validate_ready4pack_manifest}
\title{Validate ready4pack manifest ready4 s3 class Manifest for packages containing datasets.}
\usage{
validate_ready4pack_manifest(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 s3 class Manifest for packages containing datasets.}
}
\value{
A prototpe for ready4 s3 class Manifest for packages containing datasets.
}
\description{
Validate an instance of the ready4 s3 class Manifest for packages containing datasets.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
