% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_metamorphose.R
\name{metamorphose.ready4pack_manifest}
\alias{metamorphose.ready4pack_manifest}
\alias{metamorphose,ready4pack_manifest-method}
\title{Metamorphose - a method that metamorphoses an instance of a class into an instance of a different (non-child) class}
\usage{
metamorphose.ready4pack_manifest(x)

\S4method{metamorphose}{ready4pack_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class Manifest for packages containing datasets.}
}
\value{
Manifest (a ready4 S3)
}
\description{
metamorphose.ready4pack_manifest() is a metamorphose method that metamorphoses an instance of a class into an instance of a different (non-child) class. This method is implemented for the ready4 s3 class Manifest for packages containing datasets. The function returns Manifest (a ready4 S3).
}
