% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4pack_manifest.R
\name{ready4pack_manifest}
\alias{ready4pack_manifest}
\title{ready4 s3 class Manifest for packages containing datasets.}
\usage{
ready4pack_manifest(x = make_pt_ready4pack_manifest())
}
\arguments{
\item{x}{A prototype for the ready4 s3 class Manifest for packages containing datasets., Default: make_pt_ready4pack_manifest()}
}
\value{
A validated instance of the ready4 s3 class Manifest for packages containing datasets.
}
\description{
Create a new valid instance of the ready4 s3 class Manifest for packages containing datasets.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
